+++ src/rdb_protocol/btree.cc
@@ -405,13 +405,13 @@ void rdb_rget_slice(btree_slice_t *slice, const key_range_t &range,
     }
 }
 
-void rdb_distribution_get(btree_slice_t *slice, int max_depth, const store_key_t &left_key, 
+void rdb_distribution_get(btree_slice_t *slice, int max_depth, const store_key_t &left_key,
                           transaction_t *txn, superblock_t *superblock, distribution_read_response_t *response) {
-    int key_count_out;
+    int64_t key_count_out;
     std::vector<store_key_t> key_splits;
     get_btree_key_distribution(slice, txn, superblock, max_depth, &key_count_out, &key_splits);
 
-    int keys_per_bucket;
+    int64_t keys_per_bucket;
     if (key_splits.size() == 0) {
         keys_per_bucket = key_count_out;
     } else  {
